requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# OAuth und Social-Login
httpx==0.25.2
PyJWT==2.8.0
cryptography==41.0.8

# Multi-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.1.0

# E-Mail und Notifications
aiosmtplib==3.0.1
jinja2==3.1.2

# Security und Monitoring
bcrypt==4.1.2
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
